Feat: Separate the structs in different files to have more order
