+++ utils/functions/packages/get.go
@@ -8,27 +8,6 @@ import (
 	"github.com/go-resty/resty/v2"
 )
 
-type dataPackage struct {
-	Id          string `json:"id"`
-	Name        string `json:"name"`
-	Author      string `json:"author"`
-	Url         string `json:"url"`
-	Description string `json:"description"`
-	Version     string `json:"version"`
-}
-
-type getMany struct {
-	Data    []dataPackage `json:"data"`
-	Error   bool          `json:"error"`
-	Message string        `json:"message"`
-}
-
-type getOne struct {
-	Data    dataPackage `json:"data"`
-	Error   bool        `json:"error"`
-	Message string      `json:"message"`
-}
-
 func GetSearch(key string) (getMany, error) {
 	client := resty.New()
 	resp, err := client.R().Get("https://moldy-api.herokuapp.com/api/v1/packages/search?key=" + key)
@@ -40,8 +19,8 @@ func GetSearch(key string) (getMany, error) {
 		panic(err)
 	}
 
-	if dataStruct.Data == nil {
-		return dataStruct, errors.New("Not data found")
+	if dataStruct.Error || dataStruct.Data == nil {
+		return dataStruct, errors.New(dataStruct.Message)
 	}
 	return dataStruct, nil
 }
